requirements/base.txt
Django>=5.1,<5.3
djangorestframework>=3.15,<3.17
django-filter>=24.3,<26.0
django-cors-headers>=4.6,<5.0
psycopg[binary]>=3.2,<4.0
celery[redis]>=5.4,<6.0
redis>=5.2,<7.0
gunicorn>=23.0,<24.0
whitenoise>=6.8,<7.0
Pillow>=11.1,<12.0
